remove methods that shouldn't be a part of the CRD
